requirements.txt
numpy==1.24.3
matplotlib==3.7.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
pillow==10.0.0
scipy==1.10.1
einops==0.6.1 